+++ src/indexHelpers.ts
@@ -9,6 +9,7 @@ export { default as isEqual } from "./helpers/isEqual";
 export { isBothNaN } from "./helpers/isEqual";
 export { default as isIntoView } from "./helpers/isIntoView";
 export { default as localeInfo } from "./helpers/localeInfo";
+export { default as mathSumFloat } from "./helpers/mathSumFloat";
 
 export { default as nestedProperty } from "./helpers/nestedProperty";
 export { default as observer } from "./helpers/observer";